refactor apple.ts
