+++ typescript/src/link/apple-utils.ts
@@ -1,0 +1,29 @@
+import {APIGatewayProxyEvent, APIGatewayProxyResult} from "aws-lambda";
+import {Platform} from "../models/platform";
+
+type AppleSubscription = {
+    receipt: string
+    originalTransactionId: string
+}
+
+export type AppleLinkPayload = {
+    platform: Platform.DailyEdition | Platform.Ios | Platform.IosPuzzles | Platform.IosEdition,
+    subscriptions: AppleSubscription[]
+}
+
+function deduplicate<T, U>(list: T[], selector: (item: T) => U): T[] {
+    return list.reduce<T[]>(
+        (agg, item) =>
+            agg.some((x) => selector(x) === selector(item)) ?
+                agg : agg.concat([item]), 
+        []
+    )
+}
+
+export function parseAppleLinkPayload(request: APIGatewayProxyEvent): AppleLinkPayload {
+    const parsed = JSON.parse(request.body ?? "") as AppleLinkPayload;
+    return {
+        ...parsed,
+        subscriptions: deduplicate(parsed.subscriptions, x => x.originalTransactionId)
+    }
+}
